# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Реализация онлайн библиотеки песен",
        "title": "Тестовое API",
        "contact": {}
    },
    "host": "localhost:5000",
    "basePath": "/songs",
    "paths": {
        "/songs": {
            "get": {
                "description": "Просмотр всех песен с пагинацией и фильтрацией данных",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Все песни",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Номер страницы",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Количество элементов на странице",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Название группы",
                        "name": "group",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата релиза",
                        "name": "releasedate",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/songs/change/{id}": {
            "put": {
                "description": "измение информации песни по id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Изменение песни",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID песни",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Информация о песне",
                        "name": "song",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Song"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/songs/create": {
            "post": {
                "description": "Добавление песни по получаемым параметрам в JSON",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Добавление песни",
                "parameters": [
                    {
                        "description": "Информация о песне",
                        "name": "song",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Song"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/songs/delete/{id}": {
            "delete": {
                "description": "удвление песни по параметру запроса (id)",
                "produces": [
                    "application/json"
                ],
                "summary": "Удаление песни",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID песни",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/songs/{id}": {
            "get": {
                "description": "Информация песни по id",
                "produces": [
                    "application/json"
                ],
                "summary": "Песня",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID песни",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "model.Song": {
            "type": "object",
            "properties": {
                "group": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "releaseDate": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        }
    }
}